refactor checks.rb: move reader method to a class of it's own
